Cleans up our use of the BLOCK_TYPE block tag.
Regularized the code by making sure that blocks which contain type initializers
are tagged as BLOCK_TYPE consistently.

DETAILS:

There is a cleanup pass (last clause in removeRandomJunk()) which removes all
blocks bearing the BLOCK_TYPE tag.  This cleanup should be applied to all
default type expressions supplied either as initializers for type vars in class
(and record) declarations or as default values for type arguments.
(Normalizations in scopeResolve merge these two cases into just the latter one.)

If this is not done, then AST branches which follow the pattern

    ('init' ('typeof' expr))

will be converted into

    expr

.  Since 'expr' was only being evaluated for its type, this will end up
inserting unwanted expressions into the generated C code.  

Removal of BLOCK_TYPE blocks following function resolution is a big hammer, but
at least it guarantees that those unwanted expressions are removed.  A gentler
approach would be to insert an assert in the case that BLOCK_TYPE removal is
removing a nontrivial expression, and then add cases to postFold() until the
assertion no longer fires.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20069 3a8e244f-b0f2-452b-bcba-4c88e055c3ca