Modifies atomics header files to not typedef flag.
[Reviewed by Sung]

In trying to link with some Fortran-to-C programs,
ran into a problem where the flag type is defined
in f2c.h, but it is also defined in chpl-atomics.h.

It doesn't really need to be defined in chpl-atomics.h,
so this patch tweaks the macros defining the atomic
operations so that they can use a different actual
type from the type used to generate the function
names and the atomic_zzz types; that way even though
we define atomic_flag, we do not define flag.

Also adds a test case that demonstrates the problem
that this patch fixes.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20070 3a8e244f-b0f2-452b-bcba-4c88e055c3ca