Adding a memory usage test for SSCA#2 kernels 2-4.
This test works similarly to the single locale performance test in
$CHPL_HOME/test/studies/ssca2/performance.

A precomp script reads the most recent version of SSCA2_kernels.chpl
from the release directory and inserts memory usage calls and prints
the results.  The new file is saved to SSCA2_kernels.chpl in this
directory.  One caveat is that the script uses the
PRINT_TIMING_STATISTICS conditionals to decide where to insert the
memory usage calls, and thus may miss some of the allocations (in
particular in kernel 4).

The test is compiled with the RMAT graph only and run for SCALE 8
(exact betweenness centrality).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20078 3a8e244f-b0f2-452b-bcba-4c88e055c3ca