Second iteration on the improvement of bulk comms
Mainly Alberto was working to make these contributions, so that we now
support bulk transfers for arrays that have suffered from a rank
change. For example,

- A[1..5, 1..3]=B[2,1..5, 1..3]
- A[1..5, 1..3]=B[1..5, 2, 1..3]
- A[1..5, 1..3]=B[1..5, 1..3, 2]

The trickiest case is the latter, due to, in the 2D alias of B
(resulting of the B[1..5, 1..3, 2] slicing), traversing the elements
of the right-most dimension requires a stride. The required
information is obtained from the blk variables: for B declared over
[1..5,1..3,1..3], blk is (9,3,1), but in the B[1..5,1..3,2] alias, blk
is(9,3))

Details: 

M       modules/internal/DefaultRectangular.chpl

Note this is work in progress. As a first approach we have implemented
two different functions doiBulkTransferStride (which deals with
default rectangular arrays and with BlockDist arrays
with non stridable domain) ) and doiBulkTransferStride2 (which deals
with BlockDist arrays with stridable domain, like [1..5 by 2, 1..5, 1..10 by
3]). Since the communication part is common to both functions, it has
been factored out in a doiBulkTransferStrideComm function. 

The reason behind the new doiBulkTransferStride2 called from
BlockArr.doiBulkTransferStride is the following. In
BlockArr.doiBulkTransferStride we calculate the read/written
sub-domain for each DefaultRectangular, locArr, storing the block
distributed array. In the stridable case, when slicing the
locArr[i].myElems[] with the calculated sub-domain, the resulting
alias has a blk that does not behave like the resulting blk for non
strided BlockDist arrays or DefaultRectangular arrays. Therefore the
function that calculate the "stride array" (needed in
chpl_comm_gets/puts) is slightly different.

M       modules/internal/ChapelArray.chpl

Current conditions to use bulk transfer stride
(chpl_useBulkTransferStride(a:[],b:[])) are:

  (a) destination and origin domains have the same stride values in the
        corresponding dimensions (ie. A[1..5 by 2, 1..5 by
        3]=B[11..15 by 2, 6..10 by 3])

   b- a.dom.numIndices==b.dom.numIndices.

However, regarding (a) we currently support different strides for
DefaultRectangular, so A[1..5 by 2, 1..7 by 3]=B[11..17 by 3, 6..10 by
2] actually works for non distributed arrays. Note that it is possible
to implement this with a single gets due to the number of elements
involved per dimension is the same ((1..5 by 2).length==(11..17 by
3).length, etc). This restriction arises due to strided get only
suport a single "count" array (shared for both origin and
destination). However, we are still working in getting the same
successful copy for block distributed arrays. After getting this
working for BlockDist arrays the condition will be more like:

 for i in [1..a._value.rank] do 
     if a._value.dom.dsiDim(i).lenght !=
     	b._value.dom.dsiDim(i).lenght then return false;

Regarding (b), we wonder if this is or not valid:

A[1..5,1..5,1..5]=B[1..1, 1..5, 1..5];

We first thought that this was legal and that one of the matrices of B
(25 elements total) should be copied into the first 25 positions of
A. Therefore, we did our best to support this (which actually works for
DefaultRectangular arrays).

Then we have realized that if we translate this op= into the
corresponding zippered iteration:

for (a,b) in (A[1..5,1..5,1..5], B[1..1,1..5,1..5]) do a=b; 

this is not legal, so we have included the
a.dom.numIndices==b.dom.numIndices restriction to filter these
cases. Note that this (b) condition is actually embedded in the future
version of the (a) condition mentioned above. Therefore this (b)
condition is poised on the brink of extinction.

M       modules/dists/BlockDist.chpl

Was not working for 1D dimensions because we were testing only 2D and
3D cases!!. Now it is working. 

M      test/users/alberto/rafatest2.chpl
A       test/users/alberto/rafatest2.good
M       test/users/alberto/2dStrideBlock.chpl
M       test/users/alberto/perfTest.chpl
M       test/users/alberto/2dStrideBlock.good
M       test/users/alberto/perfTest.good
M       test/users/alberto/rankChange.chpl
M       test/users/alberto/test_rank_change2.chpl
M       test/users/alberto/test_rank_change2.good

Modified test to check rank change correct behavior.

More things that we will tackle next:
     - We will change the order in with doiBulkTransfer and
     doiBulkTransferStride are called. 
     - More experimental results in HECToR
     - for (i,j) in ([1..30 by 3], [1..5 by 3, 1..10 by 2]) do
     ... is valid, so, should we support A[1..30 by 3]=B[1..5 by 3,
     1..10 by 2] (where A.rank!=B.rank)????. The thing is that several
     gets would be necessary due to for A, count=(10) but for B,
     count = (2,5) and gets only admit one "count array", so we may
     need 2 gets in this example.




git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/bulkComms@20081 3a8e244f-b0f2-452b-bcba-4c88e055c3ca