Allow running the kernel using either remote tasks or remote references.
[Reviewed by Brad.]

This adds a config variable, which controls whether the computational
kernel is done using remote tasks ("on" statements) so that the table
updates are computed locally, or direct remote references for the table
updates.  Some communication layer implementations perform better with
one, and some with the other.  For now at least, the default is to use
direct remote references with the uGNI-based communication layer, and
remote tasks with the rest.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20094 3a8e244f-b0f2-452b-bcba-4c88e055c3ca