Replace sync vars with atomics for _EndCount
[reviewed by gbt]

The _EndCount class is used to join tasks at then end of sync
statements, cobegins, and coforalls. This change replaces the two sync
variables used to perform the synchronization with a single atomic
int.  Thankfully, this also greatly simplifies the code.

I also documented the functions and how they are expected to be used.
Any changes in how we handle sync, cobegin, or coforall should be
reflected here.  FYI, I was going to put some info in the compiler
regarding this, but use of _EndCount is all over the place.  I'm just
going to assume that if one were to change anything about it, it would
be natural to look up what these _EndCount functions that were being
inserted by the compiler actually did.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20106 3a8e244f-b0f2-452b-bcba-4c88e055c3ca