Correct TEPS reporting for approximate betweenness centrality (kernel 4).
The TEPS rating we have been reporting was correct for the exact case of
betweenness centrality, in which every vertex in the graph is used as
the starting vertex for the betweenness centrality computation.  But it
was incorrect for the approximate case, in which only a subset of the
vertices are so used.  Here we fix the TEPS rating for the approximate
case.

The effect of this is that now, all other things being equal, changing
the LOW_APPROX_SCALE and TOP_APPROX_SCALE config constants from run to
run, and thus the number of starting vertices, will not appreciably
affect the reported TEPS value, at least within the range of graph sizes
for which performance scales.  The kernel 4 time will change but the
number of traversed edges will change proportionately, and the TEPS
rating will remain approximately the same.

There are no corresponding changes needed to testing comparison files,
because for regular testing we don't include TEPS ratings as part of the
compared results, and for performance testing we are computing exact
betweenness centrality, for which the new code is equivalent to the old.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20107 3a8e244f-b0f2-452b-bcba-4c88e055c3ca