Fix for the global array/domain broadcast bug
[reviewed by bradc]

The bug:

For any module-level domain or array that is declared with an
expression only (i.e., without the type), we were not properly
broadcasting the pointer to other locales.

Some background;

There was an attempted a fix at this in r15865:

      Reworked how the compiler inserts the private broadcast calls.
      This is done at the same time that the compiler puts local
      variables on the heap and relied on the removed assumption about
      the first definition for the broadcasts of arrays and domains.
      Instead of looking at the first def, the compiler now looks for
      the first use or def in the initialization function of the
      module in which the global array or domain is declared and
      places the private broadcast after this use or def.

      This fixes the 15 GASNet regressions.

I suspect this fix got the GASNet regressions working, because
previously they probably hit an assertion later in the code.  The fix
generates code that broadcasts 0 as the address, because the domain or
array has not been initialized.  If the test programs do not access
the domain or array on a locale other than the one that it is declared
on, the bug goes unnoticed.

The fix:

Currently, the way we handle initializing expressions for domains and
arrays is to create a function that does the initialization (if the
expression is not already a simple function call).  We then convert
all such functions to take the returned value as a ref argument.  The
previous fix was inserting the broadcast call after the statement that
captured the reference to the original array or domain.

This fix looks at each statement of the module initialization function
in program order and looks for defs and uses of the variable in
question (same as before).  If a def is found, the broadcast is
inserted after the statement containing the def (unchanged).  If a use
is found before finding a def, we assume that we are dealing with an
initializing expression and begin looking for the call to the
initializing function with the captured reference as an argument.
Once that call is found, the broadcast call after that statement.

Notice that the existing code works when the type is present in the
declaration and is still handled in the first (def) case above.  This
is because if the type is present, we pass a captured reference to a
temporary to the initialization function and then assign the real
domain or array to that temporary (the def).  This is a result of
replacing the default initialization for the type with the
initializing expression.

I've added comments and (more importantly) assertions that will
hopefully fire if we ever change the way we handle such initializing
expressions.

I also made isDefAndOrUse() a public (non-static) function for use in
this fix.  An alternative would be to build the full def-use maps, but
that is overkill.  If we end up needing them for something else during
makeHeapAllocations(), it might be worth doing converting this code to
use the maps.

Finally, at Brad's suggestion, I moved the code that inserts the
broadcast calls to a separate function for readability.

Testing changes:

Remove the futures that I filed for the bug.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20114 3a8e244f-b0f2-452b-bcba-4c88e055c3ca