Add on statements to associative domain functions.
[reviewed by gbt]

Adding atomics to DefaultAssociativeDom in r20109 resulted in variable
comm counts for SSCA2 kernel 2 which turned out to be due to the lock
(on clause for each try).  Bad.

This change wraps the contents of some or all of the following
functions in an on statement DefaultAssociativeDom:

  - dsiIndsIterSafeForRemoving() (after yielding indicies)
  - dsiClear()
  - dsiAdd()
  - dsiRemove()

This change reduces the overall comm counts in kernel 2
(gets/puts/forks) for all locales.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20118 3a8e244f-b0f2-452b-bcba-4c88e055c3ca