This is a small code cleanup: The code to generate the call to an initialize method is simplified if we can assume that it returns void.
Reviewed by; Vass

The original idea expressed in the code was that the initialize method could
return somthing, and the constructor would return whatever that was instead of
the newly-constructed object.  But this is contrary to what one expects from a
constructor.  Also, the feature is not used in our existing module code.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20124 3a8e244f-b0f2-452b-bcba-4c88e055c3ca