Fixes type-punned pointers in comm-gasnet.
[reviewed by Sung and Greg]

We were getting type-punned pointer warnings
(errors if CHPL_DEVELOPER is set) from
comm-gasnet.c. It turns out that the code
actually was type-punning. This patch fixes
the type-punning using memcpy. The consensus
among the Chapel developers is that fixing
the problem with a 'union' is a better
solution, but we have a tested patch for
the memcpy version... and a good compiler
should turn the memcpy into a load or store
in any case.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20129 3a8e244f-b0f2-452b-bcba-4c88e055c3ca