Steps toward LULESH unification and determinism
This change reduces the differences between our versions of lulesh
that use structured and unstructured data structures using 'diff'.
The changes involve minor things like unifying names, spacing, and
tabbing; intermediate things like compiler flags and output format; as
well as more major things like creating iterators to mimic arrays.
The goal is to take a step toward a version that can switch between
structured or unstructured data structures with a command-line flag
(though I haven't gotten there yet).

Perhaps more importantly, this change fixes the nondeterminisms in
forall loops for the original version that we caught in the release
version just prior to the 1.5 release but hadn't folded back into the
original version.  This should remove the nondeterministic failures
that we've been seeing for some time now in LULESH.

Next steps: Work on restructuring the two versions so that all
data-structure-related differences are factored into different
modules or files.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20132 3a8e244f-b0f2-452b-bcba-4c88e055c3ca