Fix filtering bug and turn on filtering for SSCA2 kernel 4
[Reviewed by gbt]

Fixed a filtering bug where edges were being added to next level even
if the edge was filtered out.  This bug has been present since we
first got the code.  This did not affect the correctness of our
nightly test runs since we had filtering turned off by default.

While I was on the topic, turned on filtering by default as required
by the SSCA #2 spec.

Updated .good files to match the results of the change.  For the comm
diagnostics test, most of the comm counts decreased.  In one case
there was a small increase in number of non-blocking forks.  This is
not unexpected as filtering edges could results in more iterations
during the forward pass.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20134 3a8e244f-b0f2-452b-bcba-4c88e055c3ca