Implement a filtering iterator for SSCA2 kernel 4
[Reviewed by gbt]

This commit adds a parallel iterator that filters edges based on
weight for the RMAT graph and the torus graph implementation used in
the nightly testing.  This iterator replaces a zippered iteration over
edges and weights, and thus avoids visiting the node (via an 'on') if
the edge is to be filtered out.  It also allows us to better hide the
graph's representation of edges and weights.

Note on logistics: The iterators are implemented in the release
version, as they will not be used and thus not affect the version
built in the release directory.  The use of the iterator is
implemented in my staging area.  This set up simplifies the build in
my staging area.

Note on performance:

For multilocale testing, The number of forks is reduced by 5-8% (due
to filtering before 'on') and the number of remote gets is reduced by
~5%.

Scalar performance is consistently better for the new version (8-18%),
but I was testing under different circumstances, so I will keep an eye
on this once it gets promoted to the release version.

Other stuff:

- Updated SSCA_kernels.chpl to match the release version before
  replacing the zippered iteration

- Updated my .compopts file to add new debug flag as per r20098

- Removed NOTEST to enable nightly testing




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20138 3a8e244f-b0f2-452b-bcba-4c88e055c3ca