Updating TRIAGE to a new format
My approach to TRIAGE is "What caused emails to go to the group this
week and how can we stamp those out to reduce noise down to zero?"  To
that end, I took a different approach to the TRIAGE file that sorts
things by issue rather than testing configuration (which is how
REGRESSIONS is organized).  An issue can be a specific benchmark that
causes problems (like lulesh) or a specific configuration that is
noisy (like qthreads).  I tried to capture everything that caused
noise this week, noting who it was assigned to (if anyone), and what
possible explorations might be.  I also sorted things into two general
categories: At the top, things that are continual pains in our side;
at the bottom things that caused noise maybe once during the week but
weren't immediately recognizable (to me anyway) as ongoing problems.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20139 3a8e244f-b0f2-452b-bcba-4c88e055c3ca