Updating 32-bit testing to run GASNet as well...
In talking about 32- vs. 64-bit locale ID issues in the LTS meeting
this week, there was a question about whether cases that I'd forgotten
to update would work worse on a 32-bit linux machine.  We haven't been
doing 32-bit linux + GASNet testing, but it seems like a good idea,
so this adds it (for example directories only).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20142 3a8e244f-b0f2-452b-bcba-4c88e055c3ca