Added size method to: array, domain, string, and range.
The size method is simply an alias to other methods in these structures (i.e. 
length, numIndices, numElements).

Adding the size method was done in an attempt to unify, in a more generic way,
the notion of "maximum number of iterations", as per an email discussion
on chapel_dev list, subject: "consistent naming", June 7-8, 2012. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20143 3a8e244f-b0f2-452b-bcba-4c88e055c3ca