Improved support for 1-tuples in Chapel
[reviewed by vass]

Prior to this change, 1-tuples could only be created in Chapel by
using the tuple() function with a single argument expression -- i.e.,
tuple(x).  This commit adds support for the syntactic construct '(x,)'
in the expression context to indicate a 1-tuple.  Note that in this
commit, the trailing comma is only permitted for 1-tuples, not for
larger tuples -- i.e., (x,y,) is a syntax error, not a 2-tuple.

Because we most recently were using blank tuple dimensions as a means
of dropping tuple components on the floor, this change also includes
switching back to using underscore for this purpose.  That is, if I
want to get the second component of a 2-tuple, I could have used:

     var ( ,x) = t2;

whereas now I have to use:

     var (_,x) = t2;

This means that _ is no longer a legal identifier by itself and is now
a parser token.

I originally tried to get 0-tuples working in this commit as well, and
got them so that they could be parsed and stored and codegened, but
they weren't all that useful/robust.  E.g., I believe I/O on them
didn't work.  My impression was that in order to implement them
robustly, we'd need to change varargs functions to accept 0 arguments
and return a 0-tuple (which may be a great idea, but was more than I
was able to tackle here).

One last-minute question that came up was that in declaration contexts,
one-tuples without dangling commas are not ambiguous.  That is, while
(x) in an expression context is simply parenthesization (and therefore
can't be considered to create a 1-tuple), in a declaration context like:

      var (x) = t1;

or:

      proc foo((x)) { ... }

The dangling comma is not required to interpret it as a 1-tuple.
After some debate, the decision was to permit these forms due to a
combination of permissiveness, a sense that it wasn't likely to be an
issue (and that we could always reconsider later if it was), and
because it's what was already implemented by the time the question
came up.


TODO: tuple() and build_tuple() are now defined identically; and it's
      no longer clear that tuple() is a useful function since its main
      value previously was to create 1-tuples.  To simplify the
      language, should we remove the user-oriented tuple() function and
      keep build_tuple() as an internal function?  Or if we want to
      keep supporting tuple(), retire build_tuple()

TODO: I need to make more spec updates to define underscore as a token
      and to describe the 1-tuple changes


File-by-file description:

added underscore to chapel.lex and refactored grammar to support
1-tuples (and, commented out, 0-tuples)
----------------------------------------------------------------
parser/chapel.lex
parser/chapel.ypp


added build routine for 1-tuples; also added initial support for
0-tuples added but commented out; refactored buildParenExpr to
pass 0- and 1-tuples through
----------------------------------------------------------------
include/build.h
AST/build.cpp


redefine _build_tuple(x...) to stop special-casing 1-tuples; the
compiler no longer calls it in this case.  Note that _build_tuple() is
now identical to tuple()
-------------------------------------------------------------------------
modules/internal/ChapelTuple.chpl


added case to handle 1-tuples in buildPromotionWrapper()
--------------------------------------------------------
resolution/wrappers.cpp


move away from using blank dimensions to drop tuple components on the floor
---------------------------------------------------------------------------
modules/standard/Sort.chpl
spec/Tuple.tex
test/*




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20146 3a8e244f-b0f2-452b-bcba-4c88e055c3ca