Remove the use of the logtmp directory (and flag)
[reviewed by bradc]

This change removes the use of the temp directory and temp file for
early logging.  Perhaps this was useful once, but it doesn't appear to
be so any more.

This was motivated by the Chapel module, where $CHPL_HOME is not
writable, and so we were losing initial information in the logfile.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20157 3a8e244f-b0f2-452b-bcba-4c88e055c3ca