Dialing down NUMRUNS from 50000 to 1000 in users/franzf/v1, users/franzf/v2.
Brad and I do not see any reason to have that many runs.
Reducing NUMRUNS should also eliminate the timeouts
that we get in a couple of testing configurations.

1000 is the default NUMRUNS in the source code,
so I simply eliminated .execopts that was setting it to 50000.
I modified the default NUMRUNS setting in the Makefiles, for consistency.

users/franzf/v0 already uses NUMRUNS=1000, so it was not affected.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20169 3a8e244f-b0f2-452b-bcba-4c88e055c3ca