Use cnselect's numcores attribute instead of the coremask one.
We use "cnselect -Lcoremask" to determine the number of cores on the
compute nodes, which we then give to the aprun "-d" option.  But the
cnselect coremask attribute is deprecated, and in any case the current
code, which is based on 32-bit ints, does not work for 32-core compute
nodes.  So, with this commit we switch from the coremask attribute to
the numcores attribute, which gives the number of cores directly.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20184 3a8e244f-b0f2-452b-bcba-4c88e055c3ca