Clarify some error messages.
The message "timedexec died with signal ..." has always been confusing
to me.  It looks for all in the world like timedexec itself has died.
But the message doesn't mean this -- timedexec prints it when the target
program has died.  Clarify this message by making it start "timedexec:
target program died with signal ...".  While I'm at it, improve a couple
of the others, too.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20199 3a8e244f-b0f2-452b-bcba-4c88e055c3ca