Fixed function resolution for sync enums
The one regression due to my enum dispatch change the other day was a David
Iten test that uses sync enums and is only tested on multiple locales
(test/multilocale/diten/needMultiLocales/DijkstraTermination.chpl).  This
commit fixes this test and adds a single-locale test of sync enum dispatch
to avoid missing this issue again in the future.

The fix is similar to what I did for enums: If, in function
resolution, we are weighing the relative value of dispatching a sync
actual to non-sync formals then we unwrap its sync-ness when
considering the relative quality of the matches.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20210 3a8e244f-b0f2-452b-bcba-4c88e055c3ca