Force single task per locale when reading the graph with graphIO in test/studies/ssca2/vass.
* To revert to possibly multiple tasks per locale, as determined by
--dataParTasksPerLocale, set --IOsingleTaskPerLocale=false.

* To force serial reading of the graph, all on Locale 0, set --IOserial=true.

* Single task per locale is forced only when the
DISTRIBUTION_TYPE param is "BLOCK".

* Writing the graph out is always serial, so far.

Single task per locale is enforced by doing a coforall over the locales
that the vertex array Row is distributed over. On each locale,
the internals of the Row array representation are examined to determine
the index set placed on that locale. The corresponding vertices are
read in serially from the appropriate parts of the files.

It would be simpler to modify the dataParTasksPerLocale setting stored
in the Row array representation (specifically, its distribution object)
before reading the graph then revert it back afterwards.
The only downside of that solution is that if the Block distribution
is modified in the future to ignore that setting, the performance may
drop without an apparent reason.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20218 3a8e244f-b0f2-452b-bcba-4c88e055c3ca