Replace several Map variables with std::map.
When building with CHPL_HOST_COMPILER=cray-xt-cray, the Map class has a problem
that sometimes causes things that were put into the map to not be there when
you try to get them back out.

In scopeResolve.cpp, I did a complete replacement, so it is no longer using Map.
In the rest of the files modified here, I only replaced Maps that I was able
trace intermittent test failures back to.

This should make the xt-cle.cray-xt.cray nightly testing more stable.

The main changes are:
  m.put(k,v);  -->  m[k] = v;
  if (T* t = m.get(k)) {}  -->  if (m.count(k) != 0) { T* t = m[k]; }
  using a std::map::iterator to walk map entries instead of copying them to
  a Vec and using forv_Vec.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20227 3a8e244f-b0f2-452b-bcba-4c88e055c3ca