Added to test/studies/graph500/v3 some improvements to Graph Construction, following today's discussion with Kristi M.
The key pieces:

* In vertex_struct, eliminated the sync var vlock$.
Instead, neighbor_count, self_edges and duplicates are now atomic.

vlock$ was used to guard the resizes of vertex_struct.Neighbors (aka
vertex_struct.nd). But since its size is determined ahead of time
(even prior to this commit), there will be no resizes while edges are
being added, so the lock is not necessary for that.

The above counts are now atomic because they can be updated concurrently.
Prior to this commit they were guarded by vlock$, I think.

* While adding neighbors, there is no longer a check whether it is
a duplicate. Instead, the intention is to eliminate the duplicates
at the end, after all edges have been added.

* Duplicate elimination is NOT implemented right now - it is a todo.

Less major:

* Converted Histogram$ elements from syncs to atomics.

* Converted a forall over indices to a zippered forall over array elements.

* Printing neighbor count histogram is guarded by the DEBUG_EDGE_HISTOGRAM
config.

* Added a mention of v3 to README.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20230 3a8e244f-b0f2-452b-bcba-4c88e055c3ca