Update to passing future in TRIAGE file
In last week's TRIAGE, Sung noted the passing future
removeEmptyRecords, which Vass had noted a few weeks prior.  Under the
argument that the TRIAGE file should read loosely in priority order
(and this isn't causing noise in the form of emails...right?), I moved
this down to a new section of things that should be addressed.  But
more importantly, I added notes on why this is currently passing, why
we haven't just retired it, and next steps.  Someone who's on triage
could/should pick this up on a slow week.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20242 3a8e244f-b0f2-452b-bcba-4c88e055c3ca