Change missing standard module from internal to user error
[reviewed in spirit by Tom and Vass]

This changes the error that occurs when a standard module
is missing from an internal error to a user error to make
it less obfuscated.  I think we were justified in making
it an internal error originally because arguably if it
was missing, something was wrong on our part (unless the
user just started going around blowing away files at random,
which is hard to intelligently defend against).  But now
that we create modules/standard/gen/.../SysCTypes.chpl at
install time, and that process could go wrong, it makes
sense to promote this to a user error as suggested by Tom
and seconded by Vass.  It doesn't affect any existing tests
(I checked) and works as expected when I blow something away
by hand.  Didn't seem worth checking in a test against to
preserve this behavior due to the delicate nature of blowing
away internal modules.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20244 3a8e244f-b0f2-452b-bcba-4c88e055c3ca