Reducing reliance on primitives and locale internals in QIO modules
[reviewed by Michael Ferguson and Casey B.]

This commit replaces a bunch of primitives in the QIO modules, primarily
on the arguments that (a) these primitives were intended for boostrapping
purposes rather than execution-time code and (b) reliance on such
internals makes the IO modules harder to maintain as the locale structure
evolves.  In general, replaced these with things like 'here' and captures
of locale values rather than storing ints as a means of encoding locales.
This also cleaned up some idioms that were a little crufty (like declaring
temps in order to determine where something was rather than 'here').

Apart from that cleanup, the main change required was to reorder the
standard modules so that Locales were set up before IO was so that IO
could refer to 'here' without causing circular module initialization
resolution issues.  This had a small effect on some tests:

* caused a writeln() in ChapelRangeBase.chpl to exhibit the same
  circularity, but this writeln() was used to generate a warning, so
  the chpl_warning primitive seems better anyway (i.e., let's not
  use writeln()s to print out warnings/errors).  Made that chnage
  and updated the .good to reflect the new formatting.

* my test that tries to replace Math.chpl with a user-supplied one
  failed for similar circularity issues due to a top-level writeln()
  statement.  Pushing it down into a function maintains the point of
  the test while dancing around this issue.

* order of modules encountered by printModStuff/foo.chpl changed as
  a result of the shuffle of module order.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20246 3a8e244f-b0f2-452b-bcba-4c88e055c3ca