New performance test for equality/inequality of identical domains
This is a simple test that times the execution time of equality/inequality
operations on identical domains.  While this is a trivial operation, it
seems likely/plausible that module code executes these kinds of checks
and that performance of them matters.

I have an optimization ready to commit that speeds up these comparisons
significantly, but am going to wait until tomorrow to commit it in order
to get a night of testing with the old implementation on the books/graphs.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20253 3a8e244f-b0f2-452b-bcba-4c88e055c3ca