Short-circuiting optimization for domain [in]equality of identical domains
[reviewed by Sung, Vass, David]

This commit makes a change to the == and != operators for domains by
having those that refer to the identical domain map class instance
return 'true'/'false' respectively as the first non-param conditional
of the routine.  This ought to fire anywhere that we rely on domain
value equality to perform optimizations internally, as well as in user
code that tests identical domains for equality/inequality.  We should
also see some significant performance improvements for the domain
equality performance test I committed Friday on the nightly
performance regressions




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20255 3a8e244f-b0f2-452b-bcba-4c88e055c3ca