Making '-cc none' the default for all aprun launchers
[reviewed by Sung, Greg]

Prior to this commit, '-cc cpu' was the default option to aprun when
CHPL_HOST_PLATFORM was unset or set to something other than 'xe-cle',
'-cc none' otherwise.  The original intention was to make it '-cc cpu'
on XTs and '-cc none' on XEs (and others) because we saw performance
beneffits on XTs for using '-cc none'.

This commit makes it '-cc none' in all cases for the following
reasons:

* The reason that '-cc cpu' would be better on XTs doesn't have a
  clear rationale, probably varies depending on the application being
  run and probably based on the tasking layer used as well.  Thus, it
  seems a little superstitious barring further investigation.

* '-cc none' seems more like what Chapel should want, and always using
  it provides uniformity.

* We have a -cc flag on the launcher that permits users who want to
  fiddle with this stuff override the default.

If we decide to reinstate this default on the XT at some future point,
we should make it depend on CHPL_TARGET_PLATFORM rather than
CHPL_HOST_PLATFORM and probably default to '-cc none' in the event of
an unset variable or unknown platform rather than the other way
around.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20257 3a8e244f-b0f2-452b-bcba-4c88e055c3ca