Update remote atomics version of my SSCA2
This version of SSCA2 uses network-supported (remote) atomics where
available.  The interface implementation is in remoteAtomics.chpl, and
it now includes support for 64-bit floating point atomics.  Config
params select the appropriate version of atomics.  This means that
when running in nightly testing, this version of SSCA2 is exactly the
same as the current release version.  Note that since our current
implementation of processor atomics does not support floating point,
we will use a sync var instead for this case.

This commit also adds support to avoid 'on' clauses in most cases when
using the remote atomics (opting for additional remote gets).
Currently I do this with a conditional expression to the 'on' clause
using a config param.  If this turns out to cause unacceptable
overhead, we can remove the 'on' or implement an optimization to
remove the on statement when we see 'on here'.

Also, picked up the change from r20187 to use an iterator over the
neighbor id and edge weight pair.

I have not enabled testing on this version yet, since it results in
the same code as the release version via config params.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20259 3a8e244f-b0f2-452b-bcba-4c88e055c3ca