Fix bug with catfiles option to sub_test
When specifying catfiles for noexec tests, the named cat files string
was not being turned into a list before passing it to the 'cat'
subprocess.  This required split() on the string (it was already being
done in the exec case).

While I was in there, I redirected stderr to stdout on the 'cat'
subprocess so that if any cat file does not exist, you would see the
error in the output.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20263 3a8e244f-b0f2-452b-bcba-4c88e055c3ca